+++ resolwe/test/__init__.py
@@ -14,10 +14,11 @@ from __future__ import absolute_import, division, print_function, unicode_litera
 
 from resolwe.test.testcases import (
     ElasticSearchTestCase, ProcessTestCase, ResolweAPITestCase, TestCase, TransactionElasticSearchTestCase,
+    TransactionTestCase,
 )
 from resolwe.test.utils import check_docker, check_installed, with_docker_executor
 
 __all__ = (
-    'TestCase', 'ProcessTestCase', 'ResolweAPITestCase', 'ElasticSearchTestCase',
+    'TestCase', 'TransactionTestCase', 'ProcessTestCase', 'ResolweAPITestCase', 'ElasticSearchTestCase',
     'TransactionElasticSearchTestCase', 'check_installed', 'check_docker', 'with_docker_executor',
 )